+++ CHANGELOG.md
@@ -1,2 +1,4 @@
+## [1.0.1] - export isolate_service
+## [1.0.0] - export isolate pool api
 ## [0.0.2] - README.md
 ## [0.0.1] - init version